Warnings in Apps removed

+++ src/applications/trafficsink/avb/CoRE4INET_AVBTrafficSinkApp.cc
@@ -62,8 +62,8 @@ void AVBTrafficSinkApp::receiveSignal(cComponent *src, simsignal_t id, cObject *
         SRPTable::TalkerEntry *tentry = check_and_cast<SRPTable::TalkerEntry*>(obj);
 
         //If talker for the desired stream, register Listener
-        if (tentry->streamId == (unsigned int) par("streamID").longValue()
-                && tentry->vlan_id == (unsigned int) par("vlan_id").longValue())
+        if (tentry->streamId == static_cast<uint64_t>(par("streamID").longValue())
+                && tentry->vlan_id == static_cast<uint16_t>(par("vlan_id").longValue()))
         {
             SRPTable *srpTable = check_and_cast<SRPTable *>(src);
 
@@ -79,8 +79,8 @@ void AVBTrafficSinkApp::receiveSignal(cComponent *src, simsignal_t id, cObject *
     else if (id == NF_AVB_LISTENER_REGISTRATION_TIMEOUT)
     {
         SRPTable::ListenerEntry *lentry = check_and_cast<SRPTable::ListenerEntry*>(obj);
-        if (lentry->streamId == (unsigned int) par("streamID").longValue()
-                && lentry->vlan_id == (unsigned int) par("vlan_id").longValue())
+        if (lentry->streamId == static_cast<uint64_t>(par("streamID").longValue())
+                && lentry->vlan_id == static_cast<uint16_t>(par("vlan_id").longValue()))
         {
             if (lentry->module == this)
             {
@@ -99,8 +99,8 @@ void AVBTrafficSinkApp::handleMessage(cMessage *msg)
 {
     if (msg && msg->isSelfMessage())
     {
-        srpTable->updateListenerWithStreamId((unsigned int) par("streamID").longValue(), this,
-                (uint16_t) par("vlan_id").longValue());
+        srpTable->updateListenerWithStreamId(static_cast<uint64_t>(par("streamID").longValue()), this,
+                static_cast<uint16_t>(par("vlan_id").longValue()));
         getDisplayString().setTagArg("i2", 0, "status/active");
         simtime_t updateInterval = par("updateInterval").doubleValue();
         if (updateInterval > 0)
@@ -131,12 +131,11 @@ void AVBTrafficSinkApp::handleParameterChange(const char* parname)
     }
     if (!parname || !strcmp(parname, "streamID"))
     {
-        this->streamID = parameterULongCheckRange(par("streamID"), 0, (unsigned long) MAX_STREAM_ID);
+        this->streamID = parameterULongCheckRange(par("streamID"), 0, MAX_STREAM_ID);
     }
     if (!parname || !strcmp(parname, "vlan_id"))
     {
-        this->vlan_id = static_cast<unsigned short>(parameterULongCheckRange(par("vlan_id"), 0,
-                (unsigned long) MAX_VLAN_ID));
+        this->vlan_id = static_cast<unsigned short>(parameterULongCheckRange(par("vlan_id"), 0, MAX_VLAN_ID));
     }
 }
